feat(jspi): enable http_request via JSPI
