+++ src/background-plugin.ts
@@ -11,17 +11,15 @@ import {
 } from './call-context.ts';
 import {
   type InternalConfig,
-  MemoryOptions,
   PluginOutput,
   SAB_BASE_OFFSET,
   SharedArrayBufferSection,
 } from './interfaces.ts';
-import { readBodyUpTo } from './utils.ts';
 import { WORKER_URL } from './worker-url.ts';
 import { Worker } from 'node:worker_threads';
 import { CAPABILITIES } from './polyfills/deno-capabilities.ts';
 import { EXTISM_ENV } from './foreground-plugin.ts';
-import { matches } from './polyfills/deno-minimatch.ts';
+import { HttpContext } from './http-context.ts'
 
 // Firefox has not yet implemented Atomics.waitAsync, but we can polyfill
 // it using a worker as a one-off.
@@ -493,95 +491,6 @@ class RingBufferWriter {
   writeFloat64(value: number): void | Promise<void> {
     this.scratchView.setFloat64(0, value, true);
     return this.write(this.scratch.slice(0, 8));
-  }
-}
-
-class HttpContext {
-  fetch: typeof fetch;
-  lastStatusCode: number;
-  lastHeaders: Record<string, string> | null;
-  allowedHosts: string[];
-  memoryOptions: MemoryOptions;
-
-  constructor(
-    _fetch: typeof fetch,
-    allowedHosts: string[],
-    memoryOptions: MemoryOptions,
-    allowResponseHeaders: boolean,
-  ) {
-    this.fetch = _fetch;
-    this.allowedHosts = allowedHosts;
-    this.lastStatusCode = 0;
-    this.memoryOptions = memoryOptions;
-    this.lastHeaders = allowResponseHeaders ? {} : null;
-  }
-
-  contribute(functions: Record<string, Record<string, any>>) {
-    functions[EXTISM_ENV] ??= {};
-    functions[EXTISM_ENV].http_request = (callContext: CallContext, reqaddr: bigint, bodyaddr: bigint) =>
-      this.makeRequest(callContext, reqaddr, bodyaddr);
-    functions[EXTISM_ENV].http_status_code = () => this.lastStatusCode;
-    functions[EXTISM_ENV].http_headers = (callContext: CallContext) => {
-      if (this.lastHeaders === null) {
-        return 0n;
-      }
-      return callContext.store(JSON.stringify(this.lastHeaders));
-    };
-  }
-
-  async makeRequest(callContext: CallContext, reqaddr: bigint, bodyaddr: bigint) {
-    if (this.lastHeaders !== null) {
-      this.lastHeaders = {};
-    }
-    this.lastStatusCode = 0;
-
-    const req = callContext.read(reqaddr);
-    if (req === null) {
-      return 0n;
-    }
-
-    const { headers, header, url: rawUrl, method: m } = req.json();
-    const method = m ?? 'GET';
-    const url = new URL(rawUrl);
-
-    const isAllowed = this.allowedHosts.some((allowedHost) => {
-      return allowedHost === url.hostname || matches(url.hostname, allowedHost);
-    });
-
-    if (!isAllowed) {
-      throw new Error(`Call error: HTTP request to "${url}" is not allowed (no allowedHosts match "${url.hostname}")`);
-    }
-
-    const body = bodyaddr === 0n || method === 'GET' || method === 'HEAD' ? null : callContext.read(bodyaddr)?.bytes();
-    const fetch = this.fetch;
-    const response = await fetch(rawUrl, {
-      headers: headers || header,
-      method,
-      ...(body ? { body: body.slice() } : {}),
-    });
-
-    this.lastStatusCode = response.status;
-
-    if (this.lastHeaders !== null) {
-      this.lastHeaders = Object.fromEntries(response.headers);
-    }
-
-    try {
-      const bytes = this.memoryOptions.maxHttpResponseBytes
-        ? await readBodyUpTo(response, this.memoryOptions.maxHttpResponseBytes)
-        : new Uint8Array(await response.arrayBuffer());
-
-      const result = callContext.store(bytes);
-
-      return result;
-    } catch (err) {
-      if (err instanceof Error) {
-        const ptr = callContext.store(new TextEncoder().encode(err.message));
-        callContext[ENV].log_error(ptr);
-        return 0n;
-      }
-      return 0n;
-    }
   }
 }
